Building stidevice and stinetwork in windows using cmake
* Using "open folder" cmake projects in visual studio
* Modified CMakeLists.txt to support MSVC build
* Several changes to get things to compile, mostly with template convert<> function definitions (seems more picky). There was a bug with RawEvent incorrectly showing up as a measurement event after conversion from TRawEvent in windows.  RawEvent::isMeasurement bool should be eliminated.
* Builds dlls for device and network library. Server builds and links and runs with these dlls.
* Including (for now) example CMakeSetting.json files for each project. These should be computer specific, but there are some subtle requirements.
* CMake can't find omniorb in windows right now (no find_package support) so I'm passing in a base path in CMakeSetting.json
* There are some issues with the install path.  Visual Studio cmake support does not allow the "installRoot" path to be the root of the CMakeLists.txt location.  So for a project in sti3/src/device/src, it will not allow the installRoot to be sti3 (it will never scan the directory on startup and cmake will not work for the project).  As a work around, I'm giving -DCMAKE_INSTALL_PREFIX as a cmakeCommandArgs.  This works, but cmake gives an error CMAKE_INSTALL_PREFIX is defined multiple times, which doesn't seem to matter.

Jason

+++ src/device/src/ShotID.cpp
@@ -23,7 +23,7 @@ ShotID ShotID::generateUniqueID(const ParseID& pid)
     std::chrono::system_clock::duration tp = now.time_since_epoch();
     std::chrono::milliseconds ms = std::chrono::duration_cast<std::chrono::milliseconds>(tp);
 
-    sid.submissionTime.timestamp = ms.count();
+    sid.submissionTime.timestamp = static_cast<double>(ms.count());
 
     if (sid.submissionTime.timestamp == lastPlayTime.timestamp) {
         //error: increment TimeStamp to ensure ShotID is unique!
